requirements.txt
click
requests
webrequests
simple-loggers
python-dateutil
